refactor: DashboardBoard, Header, SideMenu 리팩토링, 함수 분리
