+++ src/components/dashboard/edit/ManageInvitations.tsx
@@ -9,8 +9,8 @@ import Button from '@/components/common/Button';
 import AddBoxIcon from '@/components/common/Icon/AddBoxIcon';
 import ModalOverlay from '@/components/common/ModalBase/ModalOverlay';
 import Pagination from '@/components/common/Pagination';
-import FormModal from '@/components/Modal/FormModal';
-import { ConfirmModal } from '@/components/Modal';
+import FormModal from '@/components/modal/FormModal';
+import { ConfirmModal } from '@/components/modal';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { useState } from 'react';
 import axios from 'axios';
