--- src/hooks/useAuth.ts
@@ -1,5 +1,0 @@
-/**
- * @file 전역 인증 상태 관리 훅
- * @description 현재 로그인한 유저의 정보(토큰 유무, 닉네임, 프로필 등)를 전역으로 관리하고 반환합니다.
- * @note 로그인/로그아웃 함수와 함께 쓰이며, 권한이 필요한 페이지에서 유저 상태를 체크할 때 호출하세요.
- */
